# src/docs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "version": "1.0.1",
    "title": "Your pet API docs",
    "description": "Endpoints documentation for your pet API"
  },
  "consumes": ["application/json", "multipart/form-data"],
  "produces": ["application/json", "multipart/form-data"],
...
  "paths": {
...
    "/api/pets": {
      "get": {
        "summary": "Get user's pets",
        "description": "Get user's pets. Requires authentication token.",
        "tags": ["Pets"],
        "security": [{ "Bearer": [] }],
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "description": "Bearer token",
            "required": true,
            "schema": {
              "type": "string",
              "example": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIi..."
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {}
          },
          "401": {
            "description": "Not authorized",
            "content": {}
          },
          "500": {
            "description": "Server error",
            "content": {}
          }
        }
      },
      "post": {
        "summary": "Add user's pet",
        "description": "Add user's pet. Requires authentication token.",
        "tags": ["Pets"],
        "security": [{ "Bearer": [] }],
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "description": "Bearer token",
            "required": true,
            "schema": {
              "type": "string",
              "example": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIi..."
            }
          }
        ],
        "requestBody": {
          "description": "New pet's data object",
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CreatePetRequest"
              }
            }
          }
        },

        "responses": {
          "201": {
            "description": "Successful operation",
            "content": {}
          },
          "400": {
            "description": "Bad request (invalid request body)",
            "content": {}
          },
          "401": {
            "description": "Not authorized",
            "content": {}
          },
          "500": {
            "description": "Server error",
            "content": {}
          }
        }
      }
    },
    "/api/pets/{petId}": {
      "delete": {
        "summary": "Delete user's pet",
        "description": "Delete user's pets. Requires authentication token.",
        "tags": ["Pets"],
        "security": [{ "Bearer": [] }],
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "description": "Bearer token",
            "required": true,
            "schema": {
              "type": "string",
              "example": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIi..."
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {}
          },
          "401": {
            "description": "No authorization",
            "content": {}
          },
          "404": {
            "description": "Not found",
            "content": {}
          },
          "500": {
            "description": "Server error",
            "content": {}
          }
        }
      }
    },
    "/api/notices/{category}": {
      "get": {
        "tags": ["Notices"],
        "summary": "Get notices by category",
        "parameters": [
          {
            "in": "path",
            "name": "category",
            "description": "Category of the task to get",
            "required": true,
            "schema": {
              "type": "string",
              "example": "sell"
            }
          },
          {
            "in": "query",
            "name": "page",
            "description": "Current page",
            "schema": {
              "type": "integer",
              "example": "1"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "description": "limit of notices per page",
            "schema": {
              "type": "integer",
              "example": "12"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {}
          },
          "404": {
            "description": "Not found",
            "content": {}
          },
          "500": {
            "description": "Server error",
            "content": {}
          }
        }
      }
    },
...
    "/api/notices/": {
      "post": {
        "tags": ["Notices"],
        "summary": "Add user's own notice",
        "description": "Create user's notuce. Requires authentication token.",
        "security": [{ "Bearer": [] }],
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "description": "Bearer token",
            "required": true,
            "schema": {
              "type": "string",
              "example": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIi..."
            }
          }
        ],
        "requestBody": {
          "description": "Object of notice",
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserNoticeRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful operation",
            "content": {}
          },
          "400": {
            "description": "Bad request",
            "content": {}
          },
          "401": {
            "description": "No authorization",
            "content": {}
          },
          "500": {
            "description": "Server error",
            "content": {}
          }
        }
      },
      "delete": {
        "tags": ["Notices"],
        "summary": "Delete user's own notice",
        "security": [{ "Bearer": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "ID of the notice to delete from own",
            "required": true,
            "schema": {
              "type": "string",
              "example": "645e533389827094770e41f1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {}
          },
          "404": {
            "description": "Not found",
            "content": {}
          },
          "401": {
            "description": "No authorization",
            "content": {}
          },
          "500": {
            "description": "Server error",
            "content": {}
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreatePetRequest": {
        "type": "object",
        "required": ["name", "birthday", "breed"],
        "properties": {
          "photo": {
            "type": "string",
            "format": "binary",
            "description": "User's avatar"
          },
          "name": {
            "type": "string",
            "description": "Pet's name",
            "minLength": 2,
            "maxLength": 16,
            "example": "Napoleon"
          },
          "birthday": {
            "type": "string",
            "description": "Pet's birthday",
            "format": "date",
            "example": "2001-01-01"
          },
          "breed": {
            "type": "string",
            "description": "Pet's breed",
            "minLength": 2,
            "maxLength": 16,
            "example": "buldog"
          },
          "comments": {
            "type": "string",
            "description": "Comment about pet",
            "minLength": 8,
            "maxLength": 120,
            "example": "This is a good dog"
          }
        }
      },
      "CreateUserNoticeRequest": {
        "type": "object",
        "required": [
          "title",
          "category",
          "name",
          "birthday",
          "breed",
          "location",
          "sex",
          "comments",
          "photo"
        ],
        "properties": {
          "photo": {
            "type": "string",
            "format": "binary",
            "description": "User's avatar"
          },
          "title": {
            "type": "string",
            "description": "Notice title",
            "minLength": 2,
            "maxLength": 120,
            "example": "Napoleon for sale"
          },
          "category": {
            "type": "string",
            "description": "Notice category",
            "enum": ["sell", "lost-found", "for-free"],
            "default": "sell",
            "example": "sell"
          },
          "name": {
            "type": "string",
            "description": "Pet's name",
            "minLength": 2,
            "maxLength": 16,
            "example": "Napoleon"
          },
          "birthday": {
            "type": "string",
            "description": "Pet's birthday",
            "format": "date",
            "example": "2023-01-01"
          },
          "breed": {
            "type": "string",
            "description": "Pet's breed",
            "minLength": 2,
            "maxLength": 16,
            "example": "Buldog"
          },
          "location": {
            "type": "string",
            "description": "Location",
            "pattern": "/^[A-Z]{1}[a-z]{1,20}$/",
            "example": "London"
          },
          "sex": {
            "type": "string",
            "description": "Pet's sex",
            "enum": ["male", "female"],
            "default": "male",
            "example": "male"
          },
          "comments": {
            "type": "string",
            "description": "Notice comments",
            "minLength": 2,
            "maxLength": 120,
            "example": "This is a good cat"
          },
          "price": {
            "type": "number",
            "description": "Notice price",
            "example": 1
          }
        }
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}